Altered Create Function in ProductsController to now use the new model
ProductsController Create will now recieve an object made with the FullProductAndCategoriesModel class.
The Create function will then retrieve the data needed from here to create a Product and an amount of ProdCat objects.
We will then fill the necessary tables in our database with the new entries.
Currently, we add new products just fine, but the Categories aren't propperly saved from the view, so this must be fixed.
Other than that, there are no craches, or other major bugs in this current version.